Split workspace settings parity gates
